BACK-END/ai-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=1.3.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.6
httpx>=0.25.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
google-auth>=2.0.0
google-auth-oauthlib>=0.7.0
google-auth-httplib2>=0.1.0

# Dépendances pour le module News
asyncpg>=0.29.0
beautifulsoup4>=4.12.0
pgvector>=0.2.0
lxml>=4.9.0
redis>=5.0.0

# Analytics
posthog>=3.0.0